Keyman10 Work - Respond to Review
* Move classes to own files
* Move remaining windows keyboard switching code into
  WindowsKeyboardSwitchingAdapter
* Other cleanup and bug fixing
