+++ SIL.Windows.Forms.Keyboarding/SIL.Windows.Forms.Keyboarding.csproj
@@ -394,9 +394,12 @@
     <Compile Include="Windows\IWindowsLanguageProfileSink.cs" />
     <Compile Include="Windows\KeymanKeyboardAdaptor.cs" />
     <Compile Include="Windows\KeymanKeyboardDescription.cs" />
+    <Compile Include="Windows\KeymanKeyboardSwitchingAdapter.cs" />
+    <Compile Include="Windows\LayoutName.cs" />
+    <Compile Include="Windows\LegacyKeymanKeyboardSwitchingAdapter.cs" />
     <Compile Include="Windows\MsTsfInterfaces.cs" />
     <Compile Include="Windows\Win32.cs" />
-    <Compile Include="Windows\WindowsXPKeyboardSwitchingAdapter.cs" />
+    <Compile Include="Windows\WindowsKeyboardSwitchingAdapter.cs" />
     <Compile Include="Windows\WinKeyboardAdaptor.cs" />
     <Compile Include="Windows\WinKeyboardDescription.cs" />
     <Compile Include="Windows\WinKeyboardUtils.cs" />
